Corrigido problema na tabela

+++ Cufa-FW/database/migrations/2023_07_25_150005_create_formularios_table.php
@@ -30,7 +30,7 @@ return new class extends Migration
             $table->integer('telefone');
             $table->integer('celular');
             $table->string('email');
-            $table->float('rendaMensalBrutaEntrevistado');
+            $table->decimal('rendaMensalBrutaEntrevistado', 10, 2);
         });
     }
